backend/requirements.txt
# FastAPI Backend Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.0.1
psutil==5.9.6

# Core ML dependencies (matching main project)
torch==2.1.0
torchvision==0.16.0
timm==0.9.12
albumentations==1.3.1
opencv-python-headless==4.8.1.78
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.26.4
